requirements.txt
numpy
pygame
tmx
neat-python
pandas
seaborn
matplotlib
keras
deap
